feat(03-02): extend AppState with Phase 3 diff display fields
- diff_scroll: u16 → usize (supports >65535 line diffs)
- New fields: diff_lines, file_summaries, diff_mode, diff_loading, hunk_offsets, selected_file_index, hunk_cursor
- Updated scroll methods to use usize arithmetic (saturating_add/sub with usize)
- scroll_bottom for Diff now uses diff_lines.len().saturating_sub(1) instead of u16::MAX
- Added apply_git_result(), next_hunk(), prev_hunk(), jump_to_selected_file() methods
- Replaced placeholder hunk methods with real hunk_offsets-based navigation
- ui/mod.rs: cast diff_scroll as u16 for Paragraph::scroll (temporary until Plan 03)

+++ airev/src/ui/mod.rs
@@ -134,7 +134,9 @@ fn render_diff(
     frame.render_widget(block, area);
 
     let placeholder = build_diff_placeholder(theme);
-    let paragraph = Paragraph::new(placeholder).scroll((state.diff_scroll, 0));
+    // diff_scroll is usize (Plan 02+); cast to u16 for Paragraph::scroll until Plan 03
+    // replaces this with a List widget that uses usize offsets directly.
+    let paragraph = Paragraph::new(placeholder).scroll((state.diff_scroll as u16, 0));
     frame.render_widget(paragraph, inner);
 }
